More fixes for Memory Manager, and only modify lower 32 bits when modifying W GPRs

+++ app/src/main/cpp/skyline/services/serviceman.cpp
@@ -257,7 +257,7 @@ namespace skyline::service {
                             break;
 
                         case ipc::ControlCommand::QueryPointerBufferSize:
-                            response.Push<u32>(0x8000);
+                            response.Push<u16>(0x8000);
                             break;
 
                         default:
